renamed thenewapps route file
